% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{add_cls}
\alias{add_cls}
\title{Method to add an ExposomeClust to a MultiDataSet}
\usage{
add_cls(object, clsSet, ...)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet}.}

\item{...}{Arguments given to \link{add_eset} from \link{MultiDataSet}.}

\item{expoSet}{An object of class \link{ExposomeClust}.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
A \link{MultiDataSet} with the \link{ExpressionSet} added as an
independent dataset.
}
\description{
This method allows to insert an object of class \link{ExposomeClust} as an
independent dataset into an object of class \link{MultiDataSet}.
}
