% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{add_exp}
\alias{add_exp}
\title{Method to add an ExposomeSet to a MultiDataSet}
\usage{
add_exp(object, expoSet, warnings = TRUE, ...)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet}.}

\item{expoSet}{An object of class \link{ExposomeSet}.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}

\item{...}{Arguments given to \link{add_eset} from \link{MultiDataSet}.}
}
\value{
A \link{MultiDataSet} with the \link{ExpressionSet} added as an
independent dataset.
}
\description{
This method allows to insert an object of class \link{ExposomeSet} as an
independent dataset into an object of class \link{MultiDataSet}.
}
\examples{
data("exposome")
library(MultiDataSet)
md <- new("MultiDataSet")
names(md)
md <- add_exp(md, exp)
names(md)
}

