% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{add_prot}
\alias{add_prot}
\title{Method to add an ExpressionSet to a MultiDataSet tagged as
protein data container.}
\usage{
add_prot(object, expoSet, ...)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet}.}

\item{expoSet}{An object of class \link{ExpressionSet} with protein data.}

\item{...}{Arguments given to \link{add_eset} from \link{MultiDataSet}.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
A \link{MultiDataSet} with the \link{ExpressionSet} added as an
independent dataset.
}
\description{
This method allows to insert an object of class \link{ExpressionSet} as an
independent dataset into an object of class \link{MultiDataSet}. The
\link{ExpressionSet} will be tagged, in the \link{MultiDataSet} as
\code{"protein"}.
}
