% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{assocGE}
\alias{assocGE}
\alias{assocGE,MultiDataSet-rexposome}
\title{Method to perform an association study between transcriptome and exposom}
\usage{
assocGE(object, formula, select, set = "exposures", ..., sva = FALSE,
  asFactor = 5, ncores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet} containing an
\code{ExposomeSet} and an \link{ExposomeSet}.}

\item{formula}{Model (without exposure/s) to be tested.}

\item{select}{One or multiple exposures existing into the
\link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).}

\item{set}{(default \code{"exposures"}) Can take values \code{"exposures"}
of \code{"phenotype"}.}

\item{...}{...}

\item{sva}{(default \code{FALSE}) If \code{TRUE} SVAnalysis is done.}

\item{asFactor}{(default \code{5}) Maximum number of unique values to
considere an exposure 'factor'. If missing the internal value of the
\link{ExposomeSet} in the given \link{MultiDataSet} will be used as
threshold.}

\item{ncores}{(default \code{1}) Number of test performed at a time.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} information
about the process is show.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
An object of class \code{ResultSet} containing the results of the
test and the associated feature-data.
}
\description{
This function allows to perform an association study between gene
expression from microarray and the exposome. An \link{ExpresionSet} is
the object storing the gene expresion and an \link{ExposomeSet} the one
storing the exposome. Both of them needs to be encapsulated in a
\link{MultiDataSet}. The association study is perform through standard
\link{limma} pipeline. The function allows to perform multiple tests using
the argument \code{exposures}.
}
\seealso{
\link{assocME} to test the association between epigenome and
exposome, \link{assocSNP} to test the association between genome and
exposome, \link{assocPRT} to test the association between proteome and
exposome, package \link{limma} for more details, \link{plotAssociation} to
plot the results
}

