% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{assocME}
\alias{assocME}
\alias{assocME,MultiDataSet-rexposome}
\title{Method to perform an association study between epigenome and exposome}
\usage{
assocME(object, formula, select, set = "exposures", area.test = FALSE,
  method = "ls", betas = FALSE, ..., sva = FALSE, vfilter = NULL,
  ncores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet} containing an
\code{ExposomeSet} and a \link{MethylationSet}.
#' @param formula Model (without exposure/s) to be tested.}

\item{select}{One or multiple exposures existing into the
\link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).}

\item{set}{(default \code{"exposures"}) Can take values \code{"exposures"}
of \code{"phenotype"}.}

\item{area.test}{(default \code{FALSE}) If set to \code{TRUE} it tests
area instead of CpG.}

\item{method}{(default \code{"ls"}) Can takes \code{"ls"} or \code{"robust"}
and it will be the method used by \link{MEAL}/\code{limma} to test
the association.}

\item{betas}{(default \code{FALSE}) By default M-values are used. If set
to \code{TRUE} beta-vaues are used instead.}

\item{...}{Other arguments passed to \link{DAProbe} or \link{DARegion}.}

\item{sva}{(default \code{FALSE}) If set to \code{TRUE}, surrogate
variable analysis is done before calling \link{DAProbe}, or \link{DARegion},
and obtained surrogate variables are added.}

\item{vfilter}{(default \code{NULL}) Numeric number of probes used
in \link{sva} if argument \code{sva} is set to \code{TRUE}.}

\item{ncores}{(default \code{1}) nomber of analysis that will be done
at the same time (see \link{mclapply}).}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} information
about the process is show.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
An object of class \code{ResultSet} containing the results of the
test and the associated feature-data.
}
\description{
This function allows to perform epigenetic wide association studies
between methylation and expsome. A \link{MethylationSet} is the object
storing the CpGs and an \link{ExposomeSet} the one storing the exposures.
The association study relies on package \link{MEAL}. The function allows to
perform multiple tests using the argument \code{exposures}.
}
\seealso{
\link{assocGE} to test the association between transcriptome
and exposome, \link{assocSNP} to test the association between genome and
exposome, package \link{MEAL} for more details, \link{plotAssociation} to
plot the results
}

