% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{assocPRT}
\alias{assocPRT}
\alias{assocPRT,MultiDataSet-rexposome}
\title{Method to perform an association study between proteome and exposom}
\usage{
assocPRT(object, formula, exposures, ..., asFactor = 5, ncores = 1,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet} containing an
\code{ExposomeSet} and an \link{ExposomeSet}.}

\item{formula}{Model (without exposure/s) to be tested.}

\item{...}{...}

\item{asFactor}{(default \code{5}) Maximum number of unique values to
considere an exposure 'factor'. If missing the internal value of the
\link{ExposmeSet} in the given \link{MultiDataSet} will be used as
threshold.}

\item{ncores}{(default \code{1}) Number of test performed at a time.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} information
about the process is show.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}

\item{select}{One or multiple exposures existing into the
\link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).}

\item{set}{(default \code{"exposures"}) if set to \code{"phenotypes"} the
variables used in the association test are the phenotypes instead of the
exposures.}

\item{sva}{(default \code{FALSE}) if set to \code{TRUE} SVa is applied on
the model and the surrogate variables are incldue in the association test.}
}
\value{
An object of class \code{ResultSet} containing the results of the
test and the associated feature-data.
}
\description{
This function allows to perform an association study between protein
expression from microarray and the exposome. An \link{ExpresionSet} is
the object storing the protein expresion and an \link{ExposomeSet} the one
storing the exposome. Both of them needs to be encapsulated in a
\link{MultiDataSet}. The association study is perform through standard
\link{limma} pipeline. The function allows to perform multiple tests using
the argument \code{exposures}.
}
\seealso{
\link{assocME} to test the association between epigenome and
exposome, \link{assocSNP} to test the association between genome and
exposone, \link{assocGE} to test the association between gene expression
and exposome, package \link{limma} for more details,
\link{plotAssociation} to plot the results
}

