% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{assocSNP}
\alias{assocSNP}
\alias{assocSNP,MultiDataSet-rexposome}
\title{Method to perform an association study between genome and diseasome
(exposome's phenotypes)}
\usage{
assocSNP(object, design, family = "binomial", filter.callrate = 0.98,
  filter.maf = 0.05, filter.pvalHWE = 0.001, ...)
}
\arguments{
\item{object}{An object of class \link{MultiDataSet} containing an
\code{ExposomeSet} and a \link{SnpSet}.}

\item{design}{Formula to be tested at \link{snp.rhs.tests}.}

\item{family}{(default \code{"binomial"}) A string defining the generalized
linear model family. This can take \code{"binomial"}, \code{"Poisson"},
\code{"Gaussian"} or \code{"gamma"}. It is given to \link{snp.rhs.tests}.}

\item{filter.callrate}{(default \code{0.98}) Filter for SNPs call-rate. Ones
under the threshold are removed.}

\item{filter.maf}{(default \code{0.05}) Filter for minimum allele frequency
(MAF). Ones under the threshold are removed.}

\item{...}{Argument given to \link{snp.rhs.tests} from \link{snpStats}.}

\item{filter.pvalHEW}{(default \code{0.001}) Filter for Hardy–Weinberg. Ones
under the threshold are removed.}
}
\description{
This function allows to perform genome wide association studies between
SNPs and phenotypes. A \link{SnpSet} is the object storing the genotypes and
an \link{ExposomeSet} the phenotypes. The association study is performed
through \link{snp.rhs.tests} from \link{snpStats}.
}
\section{Warning}{

A difference between \code{assocSNP} and both \code{assocGE}
and \code{assocME}, is that \code{assocSNP} cannot perform multiple
test in a single run.
}
\seealso{
\link{assocGE} to test the association between transcriptome
and exposome, \link{assocME} to test the association between epigenome
and exposome, package \link{snpStats} for more details,
\link{plotAssociation} to plot the results
}

