% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{classification}
\alias{classification}
\title{Method to get the classification of the samples from an ExposomeClust.}
\usage{
classification(object)
}
\arguments{
\item{object}{An \link{ExposomeClust} to get the samples' classification.}
}
\value{
A labelled vector with the classification of each exposure.
}
\description{
Method to get the classification of the samples from an ExposomeClust.
}
\examples{
data("eclust")
tt <- classification(expo_c)
table(tt)
}
\seealso{
\link{clustering} as a constructor for \link{ExposomeClust},
\link{plotClassification} to plot the groups
}
