% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{correlation}
\alias{correlation}
\alias{correlation,ExposomeSet-methods}
\title{Creation of an ExposomeCorr from an ExposomeSet.}
\usage{
correlation(object, ..., warnings = TRUE)
}
\arguments{
\item{object}{\code{ExposomeSet} which exposures will be used to calculate
their correlation}

\item{...}{Other arguments passed to \link{cor}, \link{cramersV} or
to \link{lm}.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
\code{ExposomeCorr} with the correlation between the selected
exposures and their description
}
\description{
Method to calculate the correlation between the exposures of an
\link{ExposomeSet}. Only numerical-exposures (non categorical) will
be computed. The function creates and returns an \link{ExposomCorr}
object.
}
\examples{
data("exposome")
exp.c <- correlation(expo)
expo.c
}
\seealso{
\link{plotCorrelation} to plot the correlations of an
\link{ExposomeCorr}, \link{clustering} to see how the exposures can
cluster samples, \link{pca} to compute PCA on exposures
}
