% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eimp.R
\docType{data}
\name{eimp}
\alias{eimp}
\title{\code{list} with imputed exposures and description for testing purpouses}
\format{An object of class \code{list} of length 2.}
\usage{
data("eimp")
}
\value{
An \code{list} object.
}
\description{
\code{list} with full set of 104 exposures,
 1200 simulated samples and 4 phenotypes (asthma status,
 BMI measure, sex, and age).
}
\examples{
data("eimp")
dim(eimp[[1]])
dim(eimp[[2]])
}
\keyword{datasets}
