% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_imp.R
\docType{data}
\name{ex_imp}
\alias{ex_imp}
\title{\code{imExposomeSet} for testing purpouses}
\format{An object of class \code{imExposomeSet} of dimension 47 x 109 x 10.}
\usage{
data("ex_imp")
}
\value{
An \code{imExposomeSet} object.
}
\description{
\code{imExposomeSet} created from the raw data included in the
 package. The creation process is ilustrated in the vignette.
}
\examples{
data("ex_imp")
ex_imp
}
\keyword{datasets}
