% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclust.R
\docType{data}
\name{expo_c}
\alias{expo_c}
\title{\code{ExposomeClust} for testing purpouses}
\format{
An object of class \code{ExposomeClust} of dimension 99 x 1200 x 5.
}
\usage{
data("eclust")
}
\value{
An \code{ExposomeSet} object.
}
\description{
\code{ExposomeClust} created from an \code{ExposomeSet}
 with full set of 104 exposures, 1200 simulated samples
 and 4 phenotypes (asthma status, BMI measure, sex, and age).
 The clustering was done using \code{hclust} and
 \code{cutree} with \code{k = 3}.
}
\examples{
data("eclust")
dim(expo_c)
table(classification(expo_c))
}
\keyword{datasets}
