% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{expos}
\alias{expos}
\title{Returns the exposures matrix of an ExposomeSet.}
\usage{
expos(object)
}
\arguments{
\item{object}{An \link{ExposomeSet}.}
}
\value{
A matrix of exposures
}
\description{
Given an \link{ExposomeSet} it returns the inner matrix of exposures,
having the exposures as columns and the samples as rows.
}
\examples{
data("exposome")
expos(expo)[1:3, 1:3]
}
