% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ExposomeCorr-dim.R,
%   R/ExposomeCorr-extract.R, R/ExposomeCorr-plotCorrelation.R
\docType{class}
\name{ExposomeCorr}
\alias{ExposomeCorr}
\alias{ExposomeCorr-class}
\alias{dim,ExposomeCorr-method}
\alias{extract,ExposomeCorr-method}
\alias{plotCorrelation,ExposomeCorr-method}
\title{Class ExposomeCorr}
\usage{
\S4method{dim}{ExposomeCorr}(x)

\S4method{extract}{ExposomeCorr}(object)

\S4method{plotCorrelation}{ExposomeCorr}(x, type = c("circos", "matrix"), ...)
}
\value{
An object of class \code{ExposomeCorr}
}
\description{
Class \code{ExposomeCorr} contains a matrix of correlations between
continuos exposures calculated using \link{cor}. It also contains the
description of each exposures (\code{fData} of the original
\link{ExposomeSet}) in order to maintain a coherence
with the original source. It extends \link{eSet-class}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: Return the dimension of the internat matrix of
correlation.

\item \code{extract}: Return the raw correlation matrix

\item \code{plotCorrelation}: Draws both a matrix of circos plot of correlations
}}

\section{Slots}{

\describe{
\item{\code{assayData}}{Contains the correlation matrix
(see \link{eSet}, \link{AssayData}).}

\item{\code{featureData}}{Contains the description of the exposures including
the family where they belong (see \link{eSet}, \link{AnnotatedDataFrame}).}

\item{\code{desc.famCol}}{Name or index of the column containing the exposure's name
(in description file).}
}}

\seealso{
\code{\link{pca}} to study the behavioud between samples
and exposures in an \code{\link{ExposomeSet}}
}
