% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R
\docType{class}
\name{ExposomePCA}
\alias{ExposomePCA}
\alias{ExposomePCA-class}
\title{Class ExposomePCA}
\description{
Class \code{ExposomePCA} contains a matrix of exposures used to
compute the PCA, also a table of phenotypes and a set congtaing the
multiple results of computing the PCA.
}
\section{Slots}{

\describe{
\item{\code{pca}}{list containing all elements of the PCA}
}}
\seealso{
\link{correlation} for constructor using an
\link{ExposomeSet},
\link{ExposomeSet-class} for origin data,
\link{ExposomeClust-class} for similar structure,
\link{eSet} for main class container
}

