% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R
\docType{class}
\name{ExposomeSet}
\alias{ExposomeSet}
\alias{ExposomeSet-class}
\title{Class ExposomeSet}
\description{
Class \code{ExposomeSet} contains the exposure levels, the exposure's
description and the samples phenotype. It is the starting object for
\code{rexposome} package and extends \link{eSet}.
}
\section{Slots}{

\describe{
\item{\code{desc.famCol}}{Name of the column containing exposures' family in
description-file.}

\item{\code{assayData}}{Contains the exposures matrix with column number equal to
\code{nrow(phenoData)} (see \link{eSet}, \link{AssayData}).}

\item{\code{phenoData:}}{Contains the phenotypes or variables experimenter-supplied
(see \link{eSet}, \link{AnnotatedDataFrame}).}

\item{\code{featureData}}{Contains the description of the exposures including
the family where they belong (see \link{eSet}, \link{AnnotatedDataFrame}).}
}}

\seealso{
\code{\link{read_exposome}} for constructor from files,
\link{create_exposome} for constructor from loaded
\code{data.frames}, \link{eSet} for main class container
}
