% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{exposureNames}
\alias{exposureNames}
\title{Getter to obtain the exposures's names of an ExposomeSet or ExposomePCA}
\usage{
exposureNames(object)
}
\arguments{
\item{object}{\link{ExposomeSet} that will be queried for the exposures's
names.}
}
\value{
The name of the exposures as a character vector.
}
\description{
Getter to obtain the exposures's names of an ExposomeSet or ExposomePCA
}
\section{Warning}{

\link{exposureNames} collides with \link{featureNames} of \link{eSet}.
Although in \code{rexposome 1.0.0} both function can be used as
synonyms, this usage is discouraged and it is not assured.
}

\examples{
data("exposome")
exposureNames(expo)
}
\seealso{
\link{phenotypeNames} to get the phenotypes,
\link{familyNames} to get the families of exposures
}
