% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{extract}
\alias{extract}
\title{Raw data from \code{ExWAS}, \code{ExposomeClust} and \code{ExposomeCorr}.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy} or \code{JaccardIndexPsy}}

\item{...}{NO USED}
}
\value{
A \code{data.frame} containing the raw result from PsyGeNET or
a \code{data.frame} with the result Jaccard Index for each disease.
}
\description{
Obtain the raw data from a PsyGeNET's query  stored in a
\code{DataGeNET.Psy} object or the raw data with all the Jaccard Index for
the disease of interest of an \code{JaccardIndexPsy} object.
}
