% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ExWAS-extract.R, R/ExWAS-plotExwas.R, R/ExWAS-tef.R
\docType{class}
\name{ExWAS}
\alias{EWAS-methods}
\alias{ExWAS}
\alias{ExWAS-class}
\alias{ExWAS-methods}
\alias{extract,ExWAS-method}
\alias{plotExwas,ExWAS-method}
\alias{tef,ExWAS-method}
\title{Class ExWAS}
\usage{
\S4method{extract}{ExWAS}(object, sort = TRUE)

\S4method{plotExwas}{ExWAS}(object, color, show.effective = TRUE)

\S4method{tef}{ExWAS}(object)
}
\arguments{
\item{sort}{If \code{TRUE}, the results are ordered by pvalue.}
}
\description{
Class \code{ExWAS} obtained from \link{exwas} method of an
\link{ExposomeSet} object, contains the result of testing the association
of exposures of an \code{ExposomeSet} to its phenotypes. "ExWAS" is the
acronym of "Exposome-Wide Association  Study". The function can be applied
to one of to many phenotypes in the \code{ExposomeSet} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain the matrix of association scores

\item \code{plotExwas}: Method to plot a manhatan plot for association between
exposures and phenitypes

\item \code{tef}: Method to obtain the Threshold for effective tests (TEF)
}}
\section{Slots}{

\describe{
\item{\code{formula}}{Tested formula}

\item{\code{comparison}}{Result of performing the test to find association between
levels of exposures and phenotype.}

\item{\code{description}}{Description of the exposures used in the ExWAS.
(in description file).}
}}
\seealso{
\link{m_exwas} for constructor using an  \link{nlExWAS}
}

