% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{get_robust_sd}
\alias{get_robust_sd}
\alias{extract}
\title{Raw data from \code{ExWAS}}
\usage{
get_robust_sd(object, ...)
}
\arguments{
\item{object}{Object of class \code{ExWAS}}

\item{...}{NO USED}
}
\value{
A \code{data.frame} containing the raw result from robust SD
}
\description{
Returns internal table of robust SD of objects of class \code{ExWAS}
}
