% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/imExposomeSet-dim.R,
%   R/imExposomeSet-expo.R, R/imExposomeSet-exposureNames.R,
%   R/imExposomeSet-extract.R, R/imExposomeSet-familyNames.R,
%   R/imExposomeSet-phenotypeNames.R, R/imExposomeSet-sampleNames.R
\docType{class}
\name{imExposomeSet}
\alias{imExposomeSet}
\alias{imExposomeSet-class}
\alias{dim,imExposomeSet-method}
\alias{expos,imExposomeSet-method}
\alias{exposureNames,imExposomeSet-method}
\alias{extract,imExposomeSet-method}
\alias{familyNames,imExposomeSet-method}
\alias{phenotypeNames,imExposomeSet-method}
\alias{sampleNames,imExposomeSet-method}
\title{Class imExposomeSet}
\usage{
\S4method{dim}{imExposomeSet}(x)

\S4method{expos}{imExposomeSet}(object)

\S4method{exposureNames}{imExposomeSet}(object)

\S4method{extract}{imExposomeSet}(object, rid = -1, ...)

\S4method{familyNames}{imExposomeSet}(object, by.exposure = FALSE)

\S4method{phenotypeNames}{imExposomeSet}(object)

\S4method{sampleNames}{imExposomeSet}(object)
}
\arguments{
\item{x}{An \code{\link{imExposomeSet}} object.}

\item{rid}{Number of the imputation to be extracted}

\item{by.exposure}{If set to \code{TRUE} ir returns the family which
each exposure belongs}
}
\value{
An object of class \code{imExposomeSet}
}
\description{
Class \code{imExposomeSet} was designed to store the exposures obtained
after a multiple imputation process done using \code{mice}. The
\code{data.frame} obtained from \code{mice} has the particulatiry
to contain the columns \code{.imp} and \code{.id} joint
to phenotypes and exposures. The slots \code{assayData}, \code{featureData},
and \code{phenoData} are coordinated.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: Returns the number of exsures, samples and phenotypes.

\item \code{expos}: Returns a \code{data.frame} with exposures.

\item \code{exposureNames}: Method to obtain samples' names

\item \code{extract}: Method to extract exposures for a single imputation

\item \code{familyNames}: Getter to obtain the families's names of the
family of each exposure.

\item \code{phenotypeNames}: Getter to obtain the families's names of the
family of each exposure.

\item \code{sampleNames}: Method to obtain samples' names
}}

\section{Slots}{

\describe{
\item{\code{nimputation}}{Number of imputations done with \code{mice}.}

\item{\code{assayData}}{\code{data.frame} containing \code{.imp}, \code{.id} and
the exposures.}

\item{\code{featureData}}{\code{data.frame} containing the description of
the exposures.}

\item{\code{phenoData}}{\code{data.frame} containing \code{.imp}, \code{.id} and
the phenotypes.}
}}

\seealso{
\code{\link{load_imputed}} to create an \code{imExposomeSet}
from \code{data.frames}
}
