% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{individualNames}
\alias{individualNames}
\alias{individualNames,ExposomeSet-methods}
\title{Getter to obtain the individual names (sample names) of an ExposomeSet.}
\usage{
individualNames(object)

individualNames(object) <- value
}
\arguments{
\item{object}{\link{ExposomeSet} that will be queried for the individuals's
names.}

\item{value}{Vector of \code{character} to be assigned as individual's
names.}

\item{object}{\link{ExposomeSet} that will be queried for the individuals's
names.}
}
\value{
The name of the individuals as a character vector.
}
\description{
Getter to obtain the individual names (sample names) of an ExposomeSet.

Setter to assign individual names (sample names) of an ExposomeSet.
}
\section{Warning}{

\link{individualNames} collides with \link{sampleNames} of \link{eSet}.
Although in \code{rexposome 1.0.0} both function can be used as
synonyms, this usage is discouraged and it is not assured.


\link{individualNames<-} collides with \link{sampleNames<-} of \link{eSet}.
Although in \code{rexposome 1.0.0} both function can be used as
synonyms, this usage is discouraged and it is not assured.
}
\examples{
data("exposome")
individualNames(exp)
data("exposome")
individualNames(exp) <- paste0(individualNames(exp), "__")
}
\seealso{
\link{individualNames} to get the name of the samples,
\link{phenotypeNames} to get the phenotypes,
\link{familyNames} to get the families of exposures

\link{individualNames} to get the name of the samples,
\link{phenotypeNames} to get the phenotypes,
\link{familyNames} to get the families of exposures
}

