% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaClayton.R
\name{lambdaClayton}
\alias{lambdaClayton}
\title{Lambda Calculation for a vector of P-Values}
\usage{
lambdaClayton(x, trim = 0.5)
}
\description{
Implementation of Clayton's lambda score for a vector of P-Values
}
\author{
Juran R. González
}

