% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imputed.R
\name{load_imputed}
\alias{load_imputed}
\title{Creation of an imExposomeSet from \code{data.frames}}
\usage{
load_imputed(data, description, description.famCol = 1,
  description.expCol = 2, exposures.asFactor = 5)
}
\arguments{
\item{data}{The \code{data.frame} of both exposures and phentoypes obtained
from \code{mice}.}

\item{description}{\code{data.frame} with the description of the exposures
(relation between exposures and exposure-family).}

\item{description.famCol}{(default \code{1}) Index where the family's
name (per exposures) if found in file "description".}

\item{description.expCol}{(default \code{2}) Index where the exposure's
name if found in file "description".}

\item{exposures.asFactor}{(default \code{5}) The exposures with more
than this number of unique items will be considered as "continuous" while
the exposures with less or equal number of items will be considered as
"factor".}
}
\value{
An object of class \link{imExposomeSet}.
}
\description{
Given a \code{data.frame} from \code{code} with the multiple
imputations of both exposures and phenotypes, join with a \code{data.frame}
with exposures' description, and object of class \code{\linl{imExposomeSet}}
is created.
}
\details{
The coordination of the information is based in the columns \code{.imp} and
\ccide{.id} from the \code{data.frame} obtained from \code{mice}. The
division of exposures and phenotypes is based in description
\code{data.frame}, that are the exposures. Hence, the elements in the
main \code{data.frame} that are not in the description, are the
phentoypes.
}
\seealso{
\link{imExposomeSet} for class description
}
