% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\docType{data}
\name{me}
\alias{me}
\title{\code{data.frame} for testing purpouses}
\format{An object of class \code{data.frame} with 654 rows and 57 columns.}
\usage{
data("me")
}
\value{
A \code{data.frame}.
}
\description{
\code{data.frame} created with \code{mice} used to test and
 ilustrate the creation of \code{imExposomeSet}.
}
\examples{
data("me")
dim(me)
colnames(me)
}
\keyword{datasets}
