% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/mExWAS-extract.R,
%   R/mExWAS-plotExwas.R
\docType{class}
\name{mExWAS}
\alias{mExWAS}
\alias{mExWAS-class}
\alias{extract,mExWAS-method}
\alias{raw,mExWAS-method}
\alias{plotExwas,mExWAS-method}
\title{Class mExWAS}
\usage{
\S4method{extract}{mExWAS}(object, sort = TRUE)

\S4method{raw}{mExWAS}(object)

\S4method{plotExwas}{mExWAS}(object)
}
\arguments{
\item{sort}{NOT USED}
}
\description{
Class \code{mExWAS} obtained from \link{m_exwas} method of an
\link{ExposomeSet} object, contains the result of testing the multiple
models of exposures of an \code{ExposomeSet} to a set of given phenotypes.
"mExWAS" is the #' acronym of "Multivariate Exposome-Wide Association
Study".
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain the generated model

\item \code{raw}: Method to obtain raw model

\item \code{plotExwas}: Method to plot a heap-map with the coeficient of each
exposure
}}

\section{Slots}{

\describe{
\item{\code{ranking}}{Results obtained from the \code{mboost} package as
\code{data.frame}.}

\item{\code{result}}{klist with the fitted model and result.}

\item{\code{phenotype}}{Name of the phenotype used in the analysys.}

\item{\code{description}}{feature data from original ExposomeSet.}
}}

\seealso{
\link{exwas} for constructor using an \link{ExWAS}
}
