% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{ndim}
\alias{ndim}
\title{Number of principal components in an ExposomePCA.}
\usage{
ndim(object)
}
\arguments{
\item{object}{\code{ExposomePCA} to obtain the number of components.}
}
\value{
The number of components in the \code{ExposomePCA}.
}
\description{
Number of principal components in an ExposomePCA.
}
\examples{
data("exposome")
epca <- pca(expo[3:7, 1:100])
ndim(epca)
}
\seealso{
\link{plotPCA} to plot the PCA values of an
\link{ExposomePCA}, \link{clustering} to see how the exposures can
cluster samples, \link{correlation} to compute the correlation between
exposures
}
