% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{nl_exwas}
\alias{nl_exwas}
\title{Testing the association between an exposure and a phenotype of an
ExposomeSet using a non-lineal aproach.}
\usage{
nl_exwas(object, phenotype, bbs.df, ..., mc.cores = 1, verbose = FALSE,
  warnings = TRUE)
}
\arguments{
\item{object}{\code{ExposomeSet} that will be used for the ExWAS.}

\item{phenotype}{Target phenotype used for the study. If missing all the
phenotypes in the \link{ExposomeSet} will be used.}

\item{bbs.df}{Trace of the hat matrix for the base-learner defining the
base-learner complexity. It is passed to \code{bbs} of \code{mboost}.}

\item{...}{Arguments passed to \code{gamboost}.}

\item{mc.cores}{(default \code{1}) Number of cores to use in the analysis.}

\item{verbose}{(default \code{FALSE}) Shows messages of the process.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} warnings will
not be displayed.}
}
\value{
Returns an object of class \link{mExWAS}
}
\description{
The \code{nl_exwas} method performs an "Non-lineal Exposome-Wide
Association Study" (M-ExWAS) using the exposures in \link{ExposomeSet}
and one of its phenotype. It uses the package \code{mboost} and its
functions \code{bols}, \code{bbs} and \code{gamboost}.
}
\examples{
\dontrun{
#Being x an ExposomeSet
w1 <- nl_exwas(x, "sex")
w2 <- nl_exwas(x, "asthma")
}
}
\seealso{
\link{extract} to obtain a table with the result of the ExWAS,
\link{plotExwas} to plot the results of the ExWAS
}
