% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/nlExWAS-extract.R,
%   R/nlExWAS-plotExwas.R
\docType{class}
\name{nlExWAS}
\alias{nlExWAS}
\alias{nlExWAS-class}
\alias{extract,nlExWAS-method}
\alias{nlExWAS-methods}
\alias{plotExwas,nlExWAS-method}
\title{Class nlExWAS}
\usage{
\S4method{extract}{nlExWAS}(object, sort = TRUE)

\S4method{plotExwas}{nlExWAS}(object)
}
\arguments{
\item{object}{A \code{\link{nlExWAS}} object.}

\item{sort}{If \code{TRUE}, the results are ordered by AIC}
}
\value{
An object of class \code{nlExWAS}
}
\description{
Class \code{nlExWAS} obtained from \link{nl_exwas} method of an
\link{ExposomeSet} object, contains the result of testing the multiple
models of exposures of an \code{ExposomeSet} to a set of given phenotypes.
"nlExWAS" is the #' acronym of "Multivariate Exposome-Wide Association
Study".
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain the matrix of AIC scores

\item \code{plotExwas}: Method to raw a bar-plot with the AIC of each model
}}

\section{Slots}{

\describe{
\item{\code{ranking}}{Results obtained from the \code{mboost} package as
\code{data.frame}.}

\item{\code{models}}{Models tested in the \code{mboost} package.}

\item{\code{n}}{Number of samples (exposures) used.}

\item{\code{predictors}}{Predictors obained from \code{mboost} package.}

\item{\code{X}}{\code{data.fame} with the phenotypes used in the models.}

\item{\code{xyformula}}{Formulas used in \code{mboost} package.}
}}

