% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{phenotypeNames}
\alias{phenotypeNames}
\title{Getter to obtain the phenotype's names of an ExposomeSet or ExposomePCA.}
\usage{
phenotypeNames(object)
}
\arguments{
\item{object}{\code{ExposomeSet} that will be queried for the phenotype's
names.}
}
\value{
The name of the phenotypes as a a character vector.
}
\description{
Getter to obtain the phenotype's names of an ExposomeSet or ExposomePCA.
}
\examples{
data("exposome")
phenotypeNames(expo)
}
\seealso{
\link{exposureNames} to get the name of the exposures,
\link{familyNames} to get the families of exposures
}
