% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{plot.qq}
\alias{plot.qq}
\title{Function to draw a QQ Plot from a vector of numbers}
\usage{
\method{plot}{qq}(values)
}
\arguments{
\item{values}{Numeric vector of P.Values}
}
\value{
An object obtained from \link{ggplot}
}
\description{
Function to draw a QQ Plot from a vector of numbers
}

