% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotEffect}
\alias{plotEffect}
\title{Function to draw a plot of the pvalues stored in an \code{ExWAS} object}
\usage{
plotEffect(x, y, ...)
}
\arguments{
\item{x}{An \code{ExWAS} object which effect will be ploted.}

\item{y}{(optional) Another \code{ExWAS} object. If provded its
effects will be ploted in Y-axis.}
}
\description{
This function draws a sort of manhattan plots using the p-value of the
association of the exposures with phenotypes of an \code{ExWAS} object.
}
\seealso{
\link{exwas} as a constructor for \link{ExWAS} objects,
\link{extract} to obtain a table with the result of the ExWAS
}

