% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotExwas}
\alias{plotExwas}
\title{Function to draw a plot of the pvalues stored in an \code{ExWAS} object}
\usage{
plotExwas(object, ...)
}
\arguments{
\item{object}{An \code{ExWAS} object which p-values will be plotted.}

\item{color}{(optional) A vector of colors. The vector must have length
equal to the number of families. The vector must be names with the
name of the families.}

\item{show.effective}{(default TRUE) draws a brown line on the
threshold given by the effective number of tests.}
}
\description{
This function draws a sort of manhattan plots using the p-value of the
association of the exposures with phenotypes of an \code{ExWAS} object.
}
\seealso{
\link{exwas} as a constructor for \link{ExWAS} objects,
\link{extract} to obtain a table with the result of the ExWAS
}
