% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotHistogram}
\alias{plotHistogram}
\alias{plotHistogram,ExposomeSet-methods}
\title{It draws a histogram for each exposure in an ExposomeSet}
\usage{
plotHistogram(x, select, density = TRUE, show.trans = FALSE)
}
\arguments{
\item{x}{\link{ExposomeSet} which exposome will be plotted.}

\item{select}{Name fo the exposure to be plotted. If missing, all
exposures will be used.}

\item{density}{(default \code{TRUE}) If \code{TRUE} a density plot is
drawn overlapping the histogram.}

\item{show.trans}{(default \code{FALSE}) If set to \code{TRUE} it will draw
a panel of auxiliar plots with the continuous exposure transformed with
\code{log} and \code{sqrt}.}
}
\value{
A \code{ggplot} object.
}
\description{
It draws a grid with an histogram per exposure in an \link{ExposomeSet}.
}
\examples{
data("exposome")
plotHistogram(expo[1:3, ], select = "ldde_lip")
}
\seealso{
\link{plotFamily} to draw the profile of a family of exposures,
\link{plotMissings} to plot the missing data from an \link{ExposomeSet}
}
