% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotPCA}
\alias{plotPCA}
\title{Ploting PCA}
\usage{
plotPCA(object, set, cmpX = 1, cmpY = 2, phenotype)
}
\arguments{
\item{object}{An onbject of class \link{ExposmePCA}}

\item{set}{Group (\code{"all"}, \code{"samples"} or \code{"exposures"})
taht will be ploted.}

\item{cmpX}{(default: \code{1}) component to be placed at X axis}

\item{cmpY}{(default: \code{2}) component to be placed at Y axis}

\item{phentoype}{If \code{set} is set to \code{"samples"} can be used to
color samples by phentoype}
}
\description{
Methdo to draw a plot for PCA contained in an \link{ExposomePCA}
}
\seealso{
\link{pca} to compite PCA on an \link{ExposomeSet}, \link{plotPCA}
to plot the PCA, \link{ExposomePCA} as main class
}
