% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Function to draw a QQ Plot from a vector of numbers}
\usage{
qq_plot(values)
}
\arguments{
\item{values}{Numeric vector of P.Values}
}
\value{
An object obtained from \link{ggplot}.
}
\description{
Function to draw a QQ Plot from a vector of numbers
}
