% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{raw}
\alias{raw}
\title{Getter for raw model gnerated with mexwas}
\usage{
raw(object)
}
\value{
An object of class \code{\link{glmnet}}
}
\description{
Getter for raw model gnerated with mexwas
}
\examples{
data("exposome")
wt <- mexwas(expo[3:7, 1:100], phenotype = "asthma", family = "binomial")
raw(wt)
}
