% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exposome.R
\name{read_exposome}
\alias{read_exposome}
\title{Creation of an ExposomeSet from files}
\usage{
read_exposome(exposures, description, phenotype, sep = ",",
  na.strings = c("NA", "-", "?", " ", ""), exposures.samCol = 1,
  description.expCol = 1, description.famCol = 2, phenotype.samCol = 1,
  exposures.asFactor = 5, std.e = c("none", "normal", "robust"),
  trn.e = "none", warnings = TRUE, ...)
}
\arguments{
\item{exposures}{String with the path to the file with the matrix of
exposures.}

\item{description}{String with the path to the file with the description of
the exposures (relation between exposures and exposure-family).}

\item{phenotype}{String with the path to the file with the phenotypes of
interest.}

\item{sep}{(default \code{","}) Separator used by \code{\link{read.table}} to
load the files "exposures", "description" and "phenotype".}

\item{na.strings}{(default \code{c("NA", "-", "?", " ", "")}) Character
defining the \code{NA} values in expsome's files.}

\item{exposures.samCol}{(default \code{1}) Index where the samples' name
are found in file "exposures".}

\item{description.expCol}{(default \code{1}) Index where the exposures'
name are found in file "description".}

\item{description.famCol}{(default \code{2}) Index where the family's
name (per exposures) if found in file "description".}

\item{phenotype.samCol}{(default \code{1}) Index where the sample's name
are found in file "phenotype".}

\item{exposures.asFactor}{(default \code{5}) The exposures with more
than this number of unique items will be considered as "continuous" while
the exposures with less or equal number of items will be considered as
"factor".}

\item{std.e}{If specified, it will be the method used to standardize
the exposures. It can be set to \code{"normal"}, where mean/sd will be used,
or to \code{"robust"}, where median/mad will be used (see
\link{standardize}).}

\item{trn.e}{If specified it can contain a names vector of functions to be
applied to the exposures's \code{data.frame} (see \link{transform}).}

\item{warnings}{(default \code{TRUE}) If \code{TRUE} shows useful
information/warnings from the process of loading the exposome.}

\item{...}{Other arguments will be given to \link{transform} if \code{trn.e}
is specified.}
}
\value{
An object of class \link{ExposomeSet}.
}
\description{
Given the files that defines the exposome (measures of exposome, exposome
description and individuals phentype) it loads them and creates an
object of type \link{ExposomeSet}.
}
\details{
The rows of the exposure's file, that corresponds to samples' names, must
be the same than the phenotype's file. In the same way, the columns in
exposure's file must be the same found as rows in description file.
}
\note{
\link{ExposomeSet}'s \code{fData} will contain two inner columns called
\code{_std}, \code{_trn}, \code{_fct} and \code{_type} in order to trace the
transformations an exposure suffers and to know, at eny moment, if an
exposure is categorical or continuous. The "description" file can contains a
column called \code{type} with values \code{"factor"} and \code{"numeric"}
to speficy how an exposure needs to be understood. If given, this column
will be renamed to \code{_type}. If not given, it will be created using
\code{exposures.asFactor} value.
}
\examples{
path <- paste0(path.package("rexposome"), .Platform$file.sep, "extdata")
description <- paste0(path, .Platform$file.sep, "description.csv")
phenotype <- paste0(path, .Platform$file.sep, "phenotypes.csv")
exposures <- paste0(path, .Platform$file.sep, "exposures.csv")
exp <- read_exposome(
exposures = exposures,
description = description,
phenotype = phenotype,
exposures.samCol = 1,
description.expCol = 2,
description.famCol = 1,
phenotype.samCol = 1
)
}
\seealso{
\link{ExposomeSet} for class description,
\link{load_exposome} for constructor from loaded
\code{data.frame}s
}
