% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{restore}
\alias{restore}
\title{#' Summary of an ExposomeSet.
#'
#' Given an \link{ExposomeSet} is shows a summary for its exposures or
#' its phenotypes.
#'
#' @name Summary
#' @rdname Summary-methods
#' @aliases Summary
#' @param object code{ExposomeSet} with 'set' will be summarized.
#' @param set Set to be sumarized (\code{"exposures"} or \code{"phenotypes"}).
#' @param select Subseting of exposures of phenotypes.
#' @return A basic description of the exposures in the \code{ExposomeSet}
#' @examples
#' data("exposome")
#' Summary(expo, set = "exposures")
#' @export Summary
setGeneric("Summary", function(object, set=c("exposures", "phenotypes"),
                               select)
    standardGeneric("Summary")
)
Restore original data of an ExposomeSet.}
\usage{
restore(object)
}
\arguments{
\item{object}{On object of class \code{ExposomeSet}.}
}
\value{
An \code{ExposomeSet} with no transformations.
}
\description{
Given an \link{ExposomeSet} is restores the original data and remove
all changes (standardization, transformation, imputation) that exposures
could have.
}
\examples{
data("exposome")
restore(expo)
}
