% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\title{Class ResultSet}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \code{rexposome}. This include the
functions \link{assoc_snp}, \link{assoc_genexp}, \link{assoc_methy},
\link{assoc_prot} and  \link{crossomics}.
}
\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{class_origin}}{Character containing the class of the objects used to
create the object.}

\item{\code{names}}{Character with the names of the datasets in the
\link{MultiDataSet} used to create the object}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{sva}}{0 if no correction process was applied. 1 if SVA was applied.
process(es).}
}}

