% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{Summary}
\alias{Summary}
\title{Summary of an ExposomeSet.}
\usage{
Summary(object, set = c("exposures", "phenotypes"), select, ...)
}
\arguments{
\item{object}{code{ExposomeSet} with 'set' will be summarized.}

\item{set}{Set to be sumarized (\code{"exposures"} or \code{"phenotypes"}).}

\item{select}{Subseting of exposures of phenotypes.}
}
\description{
Given an \link{ExposomeSet} is shows a summary for its exposures or
its phenotypes.
}
\examples{
data("exposome")
Summary(exp, "exposures")
}
