% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{toES}
\alias{toES}
\title{Method to convert an imExposomeSet to an ExposomeSet}
\usage{
toES(object, rid = 1)
}
\arguments{
\item{object}{An object of class \code{imExposomeSet}}

\item{rid}{(default \code{1}) Number of the imputation to be extracted}
}
\value{
An object of class \code{\link{ExposomeSet}}-
}
\description{
This methods allows to select an imputed-set and use it to create an
\code{\link{ExposomeSet}} from an \code{\link{imExposomeSet}}.
}
\examples{
data("ex_imp")
toES(ex_imp, rid = 1)
}
