% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Function to draw a Volcano Plot}
\usage{
volcano_plot(pval, fc)
}
\arguments{
\item{pval}{numeric vector of P.Values}

\item{fc}{numeric vector of fold change}
}
\value{
A \code{ggplot} object
}
\description{
Function that takes two numeric vectors (P.Value and fold change)
and draws a volcano plot using \link{ggplot2}
}

