% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/02_plot.R,
%   R/ExposomeClust-classification.R, R/ExposomeClust-plotClassification.R,
%   R/ExposomeClust-sampleNames.R
\docType{class}
\name{ExposomeClust}
\alias{ExposomeClust}
\alias{ExposomeClust-class}
\alias{plot,ExposomeClust,ANY-method}
\alias{classification,ExposomeClust-method}
\alias{plotClassification,ExposomeClust-method}
\alias{sampleNames,ExposomeClust-method}
\title{Class ExposomeClust}
\usage{
\S4method{plot}{ExposomeClust,ANY}(x, y, ...)

\S4method{classification}{ExposomeClust}(object)

\S4method{plotClassification}{ExposomeClust}(object, type = "heatmap", ...)

\S4method{sampleNames}{ExposomeClust}(object)
}
\arguments{
\item{x}{Object of class \code{ExposomeClust}}

\item{y}{NOT USED}

\item{...}{Argument given to \code{heatmap.2}}

\item{object}{An object of class \link{ExposomeClust}}

\item{type}{(default \code{"heatmap"}) Type of plot.}
}
\value{
An object of class \code{ExposomeClust}
}
\description{
Class \code{ExposomeClust} obtained from \link{clustering} on an
\link{ExposomeSet} object, represents the groups of samples created
applying a clustering method on the \code{ExposomeSet}' exposures.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Wrapper for \code{plotClassification} method.

\item \code{classification}: Return classe asigned to each sample

\item \code{plotClassification}: Draws a heatmap for the samples' classification.

\item \code{sampleNames}: Method to obtain samples' names
}}

\section{Slots}{

\describe{
\item{\code{model}}{Result obtained on applying \code{method} on the exposures.}

\item{\code{method}}{Function used to perform the clustering of the exposures.}

\item{\code{call}}{Call used to create this object.}

\item{\code{samples}}{Name of the exposures after the clustering process.}
}}

\seealso{
\code{\link{clustering}} to apply a clustering on an
\code{\link{ExposomeSet}} and create an \code{\link{ExposomeClust}}.
}
