% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R,
%   R/ExposomeClust-classification.R, R/ExposomeClust-plotClassification.R
\docType{class}
\name{ExposomeClust}
\alias{ExposomeClust}
\alias{ExposomeClust-class}
\alias{classification,ExposomeClust-method}
\alias{plotClassification,ExposomeClust-method}
\title{Class ExposomeClust}
\usage{
\S4method{classification}{ExposomeClust}(object)

\S4method{plotClassification}{ExposomeClust}(x, type = "heatmap", ...)
}
\description{
Class \code{ExposomeClust} obtained from \link{clustering} on an
\link{ExposomeSet} object, represents the groups of samples created
applying a clustering method on the \code{ExposomeSet}' exposures.
}
\section{Methods (by generic)}{
\itemize{
\item \code{classification}: Return classe asigned to each sample

\item \code{plotClassification}: Draws a heatmap for the samples' classification.
}}

\section{Slots}{

\describe{
\item{\code{model}}{Result obtained on applying \code{method} on the exposures.}

\item{\code{method}}{Function used to perform the clustering of the exposures.}

\item{\code{call}}{Call used to create this object.}

\item{\code{samples}}{Name of the exposures after the clustering process.}
}}

\seealso{
\code{\link{clustering}} for constructor using an
\code{\link{ExposomeSet}}
}
