% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R
\docType{class}
\name{ExposomeCorr}
\alias{ExposomeCorr}
\alias{ExposomeCorr-class}
\title{Class ExposomeCorr}
\description{
Class \code{ExposomeCorr} contains a matrix of correlations between
continuos exposures calculated using \link{cor}. It also contains the
description of each exposures (\code{fData} of the original
\link{ExposomeSet}) in order to maintain a coherence
with the original source. It extends \link{eSet-class}.
}
\section{Slots}{

\describe{
\item{\code{assayData}}{Contains the correlation matrix
(see \link{eSet}, \link{AssayData}).}

\item{\code{featureData}}{Contains the description of the exposures including
the family where they belong (see \link{eSet}, \link{AnnotatedDataFrame}).}

\item{\code{desc.famCol}}{Name or index of the column containing the exposure's name
(in description file).}
}}
\seealso{
\link{correlation} for constructor using an
\link{ExposomeSet},
\link{ExposomeSet-class} for origin data,
\link{ExposomeClust-class} for similar structure,
\link{eSet} for main class container
}

