% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ExposomePCA-plotEXP.R,
%   R/ExposomePCA-plotPCA.R
\docType{class}
\name{ExposomePCA}
\alias{ExposomePCA}
\alias{ExposomePCA-class}
\alias{plotEXP,ExposomePCA-method}
\alias{plotPCA,ExposomePCA-method}
\title{Class ExposomePCA}
\usage{
\S4method{plotEXP}{ExposomePCA}(object, exposure)

\S4method{plotPCA}{ExposomePCA}(object, set, cmpX = 1, cmpY = 2, phenotype)
}
\description{
Class \code{ExposomePCA} contains a matrix of exposures used to
compute the PCA, also a table of phenotypes and a set congtaing the
multiple results of computing the PCA.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotEXP}: Plot correlation between exposures and PCA

\item \code{plotPCA}: Plot association score between phentoypes and PCA
}}

\section{Slots}{

\describe{
\item{\code{pca}}{list containing all elements of the PCA}
}}

\seealso{
\link{correlation} for constructor using an
\link{ExposomeSet},
\link{ExposomeSet-class} for origin data,
\link{ExposomeClust-class} for similar structure,
\link{eSet} for main class container
}
