% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ExposomePCA-exposureNames.R,
%   R/ExposomePCA-extract.R, R/ExposomePCA-ndim.R,
%   R/ExposomePCA-phenotypeNames.R, R/ExposomePCA-plot3PCA.R,
%   R/ExposomePCA-plotEXP.R, R/ExposomePCA-plotPCA.R, R/ExposomePCA-plotPHE.R
\docType{class}
\name{ExposomePCA}
\alias{ExposomePCA}
\alias{ExposomePCA-class}
\alias{exposureNames,ExposomePCA-method}
\alias{extract,ExposomePCA-method}
\alias{ndim,ExposomePCA-method}
\alias{phenotypeNames,ExposomePCA-method}
\alias{plot3PCA,ExposomePCA-method}
\alias{plotEXP,ExposomePCA-method}
\alias{plotPCA,ExposomePCA-method}
\alias{plotPHE,ExposomePCA-method}
\title{Class ExposomePCA}
\usage{
\S4method{exposureNames}{ExposomePCA}(object)

\S4method{extract}{ExposomePCA}(object, table = "exposures")

\S4method{ndim}{ExposomePCA}(object)

\S4method{phenotypeNames}{ExposomePCA}(object)

\S4method{plot3PCA}{ExposomePCA}(object, cmpX, cmpY, cmpZ, phenotype, main,
  angle = 35, pch = 16, legend = TRUE, plines = TRUE)

\S4method{plotEXP}{ExposomePCA}(object, exposure)

\S4method{plotPCA}{ExposomePCA}(object, set, cmpX = 1, cmpY = 2, phenotype)

\S4method{plotPHE}{ExposomePCA}(object, phenotype, exp2fac = 5)
}
\description{
Class \code{ExposomePCA} contains a matrix of exposures used to
compute the PCA, also a table of phenotypes and a set congtaing the
multiple results of computing the PCA.
}
\section{Methods (by generic)}{
\itemize{
\item \code{exposureNames}: Getter to obtain the exposures's names.

\item \code{extract}: Method to extract the raw results of the PCA.

\item \code{ndim}: Number of principal components in an ExposomePCA.

\item \code{phenotypeNames}: Getter to obtain the phenotype's names.

\item \code{plot3PCA}: Methdo to draw a 3D plot for PCA

\item \code{plotEXP}: Plot correlation between exposures and PCA

\item \code{plotPCA}: Methdo to draw a 2D plot for PCA

\item \code{plotPHE}: Plot association score between phentoypes and PCA
}}

\section{Slots}{

\describe{
\item{\code{pca}}{list containing all elements of the PCA}
}}

\seealso{
\link{correlation} for constructor using an
\link{ExposomeSet},
\link{ExposomeSet-class} for origin data,
\link{ExposomeClust-class} for similar structure,
\link{eSet} for main class container
}
