% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{extract}
\alias{extract}
\title{Raw data from \code{ExWAS}, \code{ExposomeClust} and \code{ExposomeCorr}.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{Object of class \code{ExWAS}, \code{ExposomeClust} or \code{ExposomeCorr}}

\item{...}{NO USED}
}
\value{
A \code{data.frame} containing the raw result from PsyGeNET or
a \code{data.frame} with the result Jaccard Index for each disease.
}
\description{
Returns internal table of results of objects of class \code{ExWAS},
\code{ExposomeClust} and \code{ExposomeCorr}.
}
