% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{extract}
\alias{extract}
\alias{extract,ExWAS-methods}
\title{Function to extract the results of the association and integration
studies.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{Object of class \link{ExWAS} or \link{ResultSet}.}

\item{phenotype}{Name of the phenotype to be extracted from an
\link{ExWAS}.}

\item{rid}{Name or index of the exposure to be extracted from
\link{ResultSet}.}

\item{sort}{(default \code{TRUE}) If \code{TRUE} table is ordered by
pvalue (only for \link{ResultSet} objects).}
}
\value{
A \code{data.frame} with the pvalues and other information from the
association a integration studies of the exposures and phenotypes and
exposures and omics data.
}
\description{
Method \code{extract} can be applied to \link{ExWAS} and \code{ResultSet}
objects to obtain a table with the results of the association and
integration studies implemented in \code{rexposome}.
}
\seealso{
\link{ExWAS}, \link{ResultSet}
}

