% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{extract}
\alias{extract}
\alias{extract,ExWAS-methods}
\title{Function to extract the results of the association and integration
studies.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{Object of class \link{ExWAS}.}

\item{phenotype}{Name of the phenotype to be extracted from an
\link{ExWAS}.}
}
\value{
A \code{data.frame} with the pvalues and other information from the
association a integration studies of the exposures and phenotypes and
exposures and omics data.
}
\description{
Method \code{extract} can be applied to \link{ExWAS} objects to obtain a
table with the results of the association studies.
}
\seealso{
\link{ExWAS}
}
