% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{extract}
\alias{extract}
\title{Function to extract the results of the association and integration
studies.}
\usage{
extract(object, sort = TRUE, ...)
}
\arguments{
\item{object}{Object of class \link{ExWAS}.}

\item{...}{NOT USED - ONLY FOR EXTENDED FUNCTIONALLITY}
}
\value{
A \code{data.frame} with the pvalues and other information from the
association a integration studies of the exposures and phenotypes and
exposures and omics data.
}
\description{
Method \code{extract} can be applied to \link{ExWAS} objects to obtain a
table with the results of the association studies.
}
\examples{
data(exposome)
w1 <- exwas(expo[1:5, ], asthma~1, family = "binomial")
extract(w1)
}
\seealso{
\link{ExWAS}
}
