% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ExWAS-extract.R,
%   R/ExWAS-plotEffect.R, R/ExWAS-plotExwas.R, R/ExWAS-plotVolcano.R,
%   R/ExWAS-tef.R
\docType{class}
\name{ExWAS}
\alias{ExWAS}
\alias{ExWAS-class}
\alias{extract,ExWAS-method}
\alias{plotEffect,ExWAS-method}
\alias{plotExwas,ExWAS-method}
\alias{plotVolcano,ExWAS-method}
\alias{tef,ExWAS-method}
\title{Class ExWAS}
\usage{
\S4method{extract}{ExWAS}(object, sort = TRUE, ...)

\S4method{plotEffect}{ExWAS}(x, y, select, xlab, ylab)

\S4method{plotExwas}{ExWAS}(object, ..., subtitles, color, exp.order,
  show.effective = TRUE)

\S4method{plotVolcano}{ExWAS}(x, p.value = -log10(0.001),
  show.effect = FALSE)

\S4method{tef}{ExWAS}(object)
}
\arguments{
\item{object}{An object of class \code{ExWAS}, \code{mExWAS} or \code{nlExWAS}.}

\item{sort}{If \code{TRUE}, the results are ordered by pvalue.}

\item{...}{NOT USED}

\item{x}{An \code{\link{ExWAS}} object}

\item{y}{An \code{\link{ExWAS}} object}

\item{select}{(optional) Vector with the selected exposures}

\item{xlab}{(optional) Label for x-axis}

\item{ylab}{(optional) Label for y-axis}

\item{color}{(optional) A vector of colors. The vector must have length
equal to the number of families. The vector must be names with the
name of the families.}

\item{exp.order}{(optional) Order of the exposures.}

\item{show.effective}{(default TRUE) draws a brown line on the
threshold given by the effective number of tests.}

\item{show.effect}{(default FALSE) Applyes an exponential
transformation on the effects of the exposures.}

\item{object}{An object of class \code{ExWAS}, \code{mExWAS} or \code{nlExWAS}.}

\item{color}{(optional) A vector of colors. The vector must have length
equal to the number of families. The vector must be names with the
name of the families.}

\item{exp.order}{(optional) Order of the exposures.}
}
\value{
An object of class \code{ExWAS}
}
\description{
Class \code{ExWAS} obtained from \link{exwas} method of an
\link{ExposomeSet} object, contains the result of testing the association
of exposures of an \code{ExposomeSet} to its phenotypes. "ExWAS" is the
acronym of "Exposome-Wide Association  Study". The function can be applied
to one of to many phenotypes in the \code{ExposomeSet} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain the matrix of association scores

\item \code{plotEffect}: Draws a plot with the confidence interval of each
exposure. Alows to compare two ExWAS instances.

\item \code{plotExwas}: Method to plot a manhatan plot for association between
exposures and phenitypes

\item \code{plotVolcano}: Method to plot a manhatan plot for association between
exposures and phenitypes

\item \code{tef}: Method to obtain the Threshold for effective tests (TEF)
}}

\section{Slots}{

\describe{
\item{\code{effective}}{Number containing the effective number of tests.}

\item{\code{formula}}{Tested formula.}

\item{\code{comparison}}{Result of performing the test to find association between
levels of exposures and phenotype.}

\item{\code{description}}{Description of the exposures used in the ExWAS.
(in description file).}
}}

\seealso{
\code{\link{exwas}} to perform an Exposome-Wide  Association Study
and to create an \code{\link{ExWAS}}, \code{\link{mexwas}} to perform a
Multivariate Exposome-Wide Association Study and to create a
\code{\link{mExWAS}}
}
