% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{highAndLow}
\alias{highAndLow}
\alias{highAndLow,ExposomeSet-methods}
\title{Function to convert all continuous exposures of an ExposomeSet
to categorical exposures}
\usage{
highAndLow(object, ngroups = 3, intervals = "standard", select,
  drop = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{\code{ExposomeSet} which exposures will be factored.}

\item{ngroups}{(default \code{3}) Number of intervals to be created.}

\item{intervals}{(default \code{"standard"}) If \code{"standard"} all levels
are keep. If \code{"extreme"} only the extreme levels are keep.}

\item{select}{If not set, receive the name of all exposures. It can takes a
character vector to select specific exposures.}

\item{drop}{(default \code{FALSE}) By default the original exposures are
kept, if it set to \code{TRUE} the original and continuous variables are
drop and only the factored ones are returned.}

\item{warnings}{(default \code{TRUE}) If \code{TRUE} warnings will be shown.}
}
\value{
A new \link{ExposomeSet} with the original and continuous exposures
plus the new factored exposures if \code{drop} was \code{FALSE}, otherwise
a new \link{ExposomeSet} with factored exposures.
}
\description{
This function is able to convert the continuous exposures of an
\link{ExposomeSet} to categorical exposures using the n-percentile groups
defined by \code{ngroups}. By default, all levels are kept but if
\code{intervals} is set to \code{"extreme"}, the levels between the extremes
(lowest and highest) are discarded and their values are set to \code{NA}.
}
\examples{
data("exposome")
# No drop
exp.hl <- highAndLow(exp, intervals = "extreme",
    select = c("ldde_lip", "logpb", "lcotinine"))
dim(exp.hl)
#[1]  107 1200    5
dim(exp)
#[1]  104 1200    5
# exps.hl has 107 exposures: the original 104 plus the new
#         3 factored exposures

# Drop
exp.hl <- highAndLow(exp, intervals = "extreme",
    select = c("ldde_lip", "logpb", "lcotinine"), drop=TRUE)
dim(exp.hl)
#[1]  104 1200    5
dim(exp)
#[1]  104 1200    5
# exps.hl has 104 exposures: the 3 factored exposures plus the
#         other 101 (104 - 3) original exposures
}
\seealso{
\link{sData} to chek the status of the exposures,
\link{trans} to transform the exposures,
\link{standardize} to standardize by normal or robust methods the exposures
}

