% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exposome.R
\name{load_exposome}
\alias{load_exposome}
\title{Creation of an ExposomeSet from \code{data.frames}}
\usage{
load_exposome(exposures, description, phenotype, description.famCol = 1,
  exposures.asFactor = 5, std.e = c("none", "normal", "robust"),
  trn.e = "none", warnings = TRUE, ...)
}
\arguments{
\item{exposures}{\code{data.frame} of exposures.}

\item{description}{\code{data.frame} with the description of the exposures
(relation between exposures and exposure-family).}

\item{phenotype}{\code{data.frame} with the phenotypes of interest.}

\item{description.famCol}{(default \code{1}) Index where the family's
name (per exposures) if found in file "description".}

\item{exposures.asFactor}{(default \code{5}) The exposures with more
than this number of unique items will be considered as "continuous" while
the exposures with less or equal number of items will be considered as
"factor".}

\item{std.e}{If specified, it will be the method used to standardize
the exposures. It can be set to \code{"normal"}, where mean/sd will be used,
or to \code{"robust"}, where median/mad will be used (see
\link{standardize}).}

\item{trn.e}{If specified it can contain a names vector of functions to be
applied to the exposures's \code{data.frame} (see \link{transform}).}

\item{warnings}{(default \code{TRUE}) If \code{TRUE} shows useful
information/warnings from the process of loading the exposome.}

\item{...}{Other arguments will be given to \link{transform} if \code{trn.e}
is specified.}
}
\value{
An object of class \link{ExposomeSet}.
}
\description{
Given three \code{data.frames} that defines the exposome (measures
of exposome, exposome description and individuals phentype) it loads
them and creates an object of type \link{ExposomeSet}.
}
\details{
The rows of the exposure's \code{data.frames}, that corresponds to samples'
names, must be the same than the phenotype's \code{data.frames}. In the
same way, the columns in exposure's \code{data.frames} must be the same
in description \code{data.frame}.
}
\note{
\link{ExposmeSet}'s \code{fData} will contain two inner columns called
\code{_std}, \code{_trn}, \code{_fct} and \code{_type} in order to trace the
transformations an exposure suffers and to know, at eny moment, if an
exposure is categorical or continuous. The "description" file can contains a
column called \code{type} with values \code{"factor"} and \code{"numeric"}
to speficy how an exposure needs to be understood. If given, this column
will be renamed to \code{_type}. If not given, it will be created using
\code{exposures.asFactor} value.
}
\seealso{
\link{ExposomeSet} for class description,
\link{read_exposome} for constructor from txt/csv
files.
}
