% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/mExWAS-extract.R
\docType{class}
\name{mExWAS}
\alias{mExWAS}
\alias{mExWAS-class}
\alias{extract,mExWAS-method}
\alias{mExWAS-methods}
\alias{raw,mExWAS-method}
\alias{mExWAS-methods}
\title{Class mExWAS}
\usage{
\S4method{extract}{mExWAS}(object, sort = TRUE)

\S4method{raw}{mExWAS}(object)
}
\arguments{
\item{sort}{NOT USED}
}
\description{
Class \code{mExWAS} obtained from \link{m_exwas} method of an
\link{ExposomeSet} object, contains the result of testing the multiple
models of exposures of an \code{ExposomeSet} to a set of given phenotypes.
"mExWAS" is the #' acronym of "Multivariate Exposome-Wide Association
Study".
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain the generated model

\item \code{raw}: Method to obtain raw model
}}

\section{Slots}{

\describe{
\item{\code{ranking}}{Results obtained from the \code{mboost} package as
\code{data.frame}.}

\item{\code{method}}{Name of the method used to evaluate the data.}

\item{\code{mresult}}{Result obtained from evaluating the model.}

\item{\code{arguments}}{Arguments passed to \code{glmnet} or to \code{partDSA}.}
}}

\seealso{
\link{exwas} for constructor using an \link{ExWAS}
}
