% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotClassification}
\alias{plotClassification}
\alias{plotClassification,ExposomeClust-methods}
\title{Draw the profile of the levels exposures after a classification with a
clustering method}
\usage{
plotClassification(x, ...)
}
\arguments{
\item{x}{Object of class \code{Exposomeclust}}

\item{...}{NOT USED}

\item{type}{Two types are available: \code{"heatmap"} or \code{"valuemap"}.}
}
\value{
A \code{list} with different graphics parameters.
}
\description{
Draw the profile of the levels exposures after a classification with a
clustering method
}
\examples{
\dontrun{
data("eclust")
plotClassification(expo_c)
}
}
\seealso{
\link{clustering} as a constructor for \link{ExposomeClust},
\link{classification} to see how to obtain the classification of
the samples from an \link{ExposomeClust}
}
