% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotEXP}
\alias{plotEXP}
\title{Plot correlation between exposures and PCA}
\usage{
plotEXP(object, exposure)
}
\arguments{
\item{object}{An object oc class \link{ExposmePCA}}

\item{exposure}{(optional) to select a set of exposures to be ploted.
If not given all are used.}
}
\description{
Function used to plot the correlation between the exposures in an
\link{ExposomePCA} and the values for each component of the PCA in the
same \link{ExposomePCa}
}
\seealso{
\link{pca} to compute PCA on an \link{ExposomeSet}, \link{plotPHE}
to plot the P-Value of association between phenotypes ans PCA,
\link{ExposmePCA} as main class
}
