% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{restore}
\alias{restore}
\title{Restore original data of an ExposomeSet.}
\usage{
restore(object)
}
\arguments{
\item{object}{On object of class \code{ExposomeSet}.}
}
\value{
An \code{ExposomeSet} with no transformations.
}
\description{
Given an \link{ExposomeSet} is restores the original data and remove
all changes (standardization, transformation, imputation) that exposures
could have.
}
\examples{
data("exposome")
restore(expo)
}
